requirements.txt
requests
python-dotenv
geoip2
pandas
numpy
scikit-learn
requests
python-dotenv
geoip2
pandas
numpy
scikit-learn
